Replace in-memory expense storage with JPA-based persistence

+++ kakeibo/src/main/java/com/example/kakeibo/exception/ExpenseNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.kakeibo.exception;
 
 public class ExpenseNotFoundException extends RuntimeException {
 
-    public ExpenseNotFoundException(int id) {
+    public ExpenseNotFoundException(Long id) {
         super("Expense not found: id=" + id);
     }
 }
